segregate connectwallet button in plotheader.vue
