+++ src/components/content/Plotheader.vue
@@ -37,24 +37,7 @@
         </el-menu>
       </el-col>
       <el-col id="connect" :lg="3" :md="5" :sm="4" :xs="5">
-        <el-button
-          v-if="!baddr"
-          @click="connect_wallet"
-          class="connect"
-          :loading="connect_loading"
-          >{{ $t("connect") }}</el-button
-        >
-        <span v-else style="color: #fff" class="baddr font">
-          <el-tooltip effect="light" placement="bottom">
-            <span slot="content" class="font">
-              {{ $t("bsc") }}: {{ baddr }}
-            </span>
-            <el-button class="font">
-              {{ baddr.substr(0, 6) + "..." + baddr.substr(-4, 4) }}
-              <span v-if="testnet">{{ testnet }}</span>
-            </el-button>
-          </el-tooltip>
-        </span>
+        <ConnectButton />
       </el-col>
       <el-col id="changelang" :lg="3" :md="5" :sm="5" :xs="5">
         <el-select v-model="lang">
@@ -106,9 +89,7 @@
 </template>
 
 <script>
-import Web3Modal from "web3modal";
-import WalletConnectProvider from "@walletconnect/web3-provider";
-
+import ConnectButton from "../lib/ConnectWalletButton.vue";
 import { mapState } from "vuex";
 import market from "../../market";
 import keeper from "../../keeper";
@@ -142,6 +123,9 @@ function tags(path) {
 }
 export default {
   name: "Plotheader",
+  components: {
+    ConnectButton,
+  },
   computed: mapState({
     baddr: "baddr",
     myList: "myList",
@@ -183,7 +167,6 @@ export default {
   },
   data() {
     return {
-      connect_loading: false,
       langs: [
         { value: "en", label: "English" },
         { value: "zh", label: "简体中文" },
@@ -198,41 +181,7 @@ export default {
       this.$store.commit("setCurrentPbtId", false);
     },
 
-    connect_wallet: async function () {
-      this.connect_loading = true;
-      const commit = this.$store.commit;
-      const providerOptions = {
-          walletconnect: {
-              package: WalletConnectProvider,
-              options: {
-                rpc: {
-                    0x38: 'https://bsc-dataseed.binance.org',
-                    0x61: 'https://data-seed-prebsc-1-s1.binance.org:8545/'
-                }
-              }
-          }
-      };
-      const wmod = new Web3Modal({
-          network: "bnb",
-          cacheProvider: true,
-          providerOptions
-      });
-      const wm_instance = await wmod.connect();
-      console.log('wm_instance', wm_instance);
-      const bsc = await market.connect(commit);
-      if (typeof bsc == "string" || !bsc) {
-        if (bsc) {
-          this.$message.error(`Connect failed: ${bsc}`);
-        } else {
-          this.connect_faild = true;
-        }
-        this.connect_loading = false;
-      } else {
-        commit("setBaddr", bsc.addr);
-        keeper.startKeeper(bsc, commit);
-        this.connect_loading = false;
-      }
-    },
+    
   },
 };
 </script>
